Added 'Cancel' button when sending contact details, fixed magic numbers in saveAdopter(), fixed possible NPE for stage1 buttons.

+++ src/main/java/pro/sky/telegramcatdog/controller/AdopterController.java
@@ -34,7 +34,7 @@ public class AdopterController {
             tags = "Работа с усыновителями"
     )
     @GetMapping("{id}")
-    public Adopter readAdopterById(@Parameter(description = "id усыновителя", example = "1") @PathVariable Integer id) {
+    public Adopter readAdopterById(@Parameter(description = "id усыновителя", example = "1") @PathVariable Long id) {
         return adopterService.readAdopter(id);
     }
